add api reference component and routing
