+++ gatsby-node.js
@@ -1,14 +1,12 @@
 const path = require('path')
 
-exports.createPages = ({
-  actions,
-  graphql
-}) => {
-  const {
-    createPage
-  } = actions
+exports.createPages = ({ actions, graphql }) => {
+  const { createPage } = actions
 
-  const docTemplate = path.resolve(`src/templates/doc-template.tsx`)
+  const docTemplate = path.resolve(`src/components/Templates/DocTemplate.tsx`)
+  const apiReferenceDoc = path.resolve(
+    `src/components/Templates/ApiReference.tsx`
+  )
 
   return graphql(`
     {
@@ -38,6 +36,10 @@ exports.createPages = ({
         component: docTemplate,
         context: {}, // additional data can be passed via context
       })
+    })
+    createPage({
+      path: '/api-reference',
+      component: apiReferenceDoc,
     })
   })
 }
