+++ package.json
@@ -10,6 +10,7 @@
     "@material-ui/styles": "^4.1.1",
     "@mdx-js/mdx": "^1.0.21",
     "@mdx-js/react": "^1.0.21",
+    "axios": "^0.19.0",
     "classnames": "^2.2.6",
     "gatsby": "^2.13.38",
     "gatsby-image": "^2.2.0",
