+++ src/components/App/App.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { Router } from '@reach/router'
 import Layout from '../Layout/Layout'
 import Main from '../Layout/Main'
+import ApiReference from '../Layout/ApiReference'
 import { MuiThemeProvider } from '@material-ui/core/styles'
 import ORDERCLOUD_THEME from '../../theme/theme.constants'
 import Route from '../Shared/Route'
@@ -14,6 +15,7 @@ class App extends React.Component {
         <Layout>
           <Router>
             <Route path={withPrefix('/')} component={<Main />} />
+            <Route path={withPrefix('/')} component={<ApiReference />} />
           </Router>
         </Layout>
       </MuiThemeProvider>
